Move NameDatabase
